Move suppressCertificateWarning to tests/config/utils
